+++ src/main/java/kr/ac/sejong/ds/palette/common/config/JpaAuditingConfiguration.java
@@ -1,4 +1,4 @@
-package kr.ac.sejong.ds.palette.config;
+package kr.ac.sejong.ds.palette.common.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
