+++ src/main/java/kr/ac/sejong/ds/palette/common/entity/BaseEntity.java
@@ -1,4 +1,4 @@
-package kr.ac.sejong.ds.palette.common;
+package kr.ac.sejong.ds.palette.common.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
